Clang: finally enable `-Wshorten-64-to-32` (for non-32-bit builds...)

+++ src/librawspeed/common/RawImage.cpp
@@ -20,6 +20,7 @@
 
 #include "rawspeedconfig.h"
 #include "common/RawImage.h"
+#include "adt/Casts.h"
 #include "adt/CroppedArray2DRef.h"
 #include "adt/Mutex.h"
 #include "adt/Point.h"
@@ -75,7 +76,8 @@ void RawImageData::createData() {
     ThrowRDE("Duplicate data allocation in createData.");
 
   // want each line to start at 16-byte aligned address
-  pitch = roundUp(static_cast<size_t>(dim.x) * bpp, alignment);
+  pitch =
+      implicit_cast<int>(roundUp(static_cast<size_t>(dim.x) * bpp, alignment));
   assert(isAligned(pitch, alignment));
 
 #if defined(DEBUG) || __has_feature(address_sanitizer) ||                      \
@@ -200,7 +202,8 @@ void RawImageData::subFrame(iRectangle2D crop) {
 void RawImageData::createBadPixelMap() {
   if (!isAllocated())
     ThrowRDE("(internal) Bad pixel map cannot be allocated before image.");
-  mBadPixelMapPitch = roundUp(roundUpDivision(uncropped_dim.x, 8), 16);
+  mBadPixelMapPitch =
+      implicit_cast<uint32_t>(roundUp(roundUpDivision(uncropped_dim.x, 8), 16));
   assert(mBadPixelMap.empty());
   mBadPixelMap.resize(static_cast<size_t>(mBadPixelMapPitch) * uncropped_dim.y,
                       uint8_t(0));
